dynamic validation

+++ src/controller/userController.js
@@ -4,8 +4,8 @@ import validateUserInput  from "../utils/validateUserInput.js";
 export const getAllUser = async (req, res) => {
   try{
     const users = await userService.getAllUserService();
-    if(!users) return res.status(404).json({message:"No Users Available ", success:false});
-    res.status(200).json({data: users, success:true});
+    if(!users) return res.status(404).json({message:"No Users Available "});
+    res.status(200).json({data: users});
   }catch(error){
     console.error('Error fetching user:', error);
     res.status(500).json({ error: 'Internal server error' });
@@ -19,8 +19,8 @@ export const getUserById = async (req, res) => {
   }
   try{
     const user = await userService.getUserByIdService(userId);
-    if (!user) return res.status(404).json({message:"User Not Found", success:false});
-    res.status(200).json({data: user, success:true});
+    if (!user) return res.status(404).json({message:"User Not Found"});
+    res.status(200).json({data: user});
   }catch(error){
     console.error('Error fetching user:', error);
     res.status(500).json({ error: 'Internal server error' });
@@ -28,33 +28,43 @@ export const getUserById = async (req, res) => {
   
 };
 
-export const createUser = async (req,res) => {
+export const signup = async (req,res) => {
+
   const {valid,message,sanitizedData} = await validateUserInput(req.body);
-  
   if(!valid){
     return res.status(400).json({error: message});
   }
+
   try{
     const newUser = await userService.createUserService(sanitizedData);
     res.status(201).json({message:'User Created', data:newUser});
   }catch(error){
     console.error('Error creating user:', error);
-    res.status(500).json({ error: 'Something went wrong' });
+    res.status(500).json({ error: 'Something went wrong'});
   }
 };
 
 export const deleteUser = async (req, res) => {
   const userId = parseInt(req.params.id);
+  if(isNaN(userId)){
+    return res.status(400).json({ message: "Invalid user ID"});
+  }
   const deletedUser = await userService.deleteUserService(userId);
   res.status(200).json({ error: 'User Deleted', data:delUser });
 };
 
 export const updateUser = async (req, res) => {
-  const userData = req.body;
   const userId = parseInt(req.params.id);
+  if(isNaN(userId)){
+    return res.status(400).json({ message: "Invalid user ID"});
+  }
 
+  const {valid,message,validatedData} = await validateUserInput(req.body,userId);
+  if(!valid){
+    return res.status(400).json({error: message});
+  }
   try{
-    const updatedUserData = await userService.updateUserService(userData,userId);
+    const updatedUserData = await userService.updateUserService(validatedData,userId);
     res.status(200).json({message:"User Updated", data:updatedUserData});
   }catch(error){
     console.log(error);
@@ -65,7 +75,7 @@ export const updateUser = async (req, res) => {
 export default {
     getAllUser,
     getUserById,
-    createUser,
+    signup,
     deleteUser,
     updateUser
 };
